local mg works for Q1
git-svn-id: https://svn.dealii.org/trunk@2761 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/include/multigrid/multigrid.h
@@ -221,7 +221,7 @@ class MGTransferPrebuilt : public MGTransferBase
 					  * while row indices belong to the
 					  * child cell, i.e. the fine level.
 					  */
-    vector<SparseMatrix<float> > prolongation_matrices;
+    vector<SparseMatrix<double> > prolongation_matrices;
 };
